+++ richat/build.rs
@@ -68,6 +68,17 @@ fn generate_grpc_geyser() -> anyhow::Result<()> {
         )
         .method(
             Method::builder()
+                .name("subscribe_accounts")
+                .route_name("SubscribeAccounts")
+                .input_type("richat_proto::richat::SubscribeAccountsRequest")
+                .output_type("Vec<u8>")
+                .codec_path("richat_shared::transports::grpc::SubscribeCodec")
+                .client_streaming()
+                .server_streaming()
+                .build(),
+        )
+        .method(
+            Method::builder()
                 .name("subscribe_replay_info")
                 .route_name("SubscribeReplayInfo")
                 .input_type("richat_proto::geyser::SubscribeReplayInfoRequest")